+++ pages/dashboard.tsx
@@ -1,16 +1,242 @@
-import React from 'react';
+/* This example requires Tailwind CSS v2.0+ */
+import { Dialog, Transition } from '@headlessui/react';
+import {
+  Bars3Icon,
+  ChartBarIcon,
+  CloudArrowUpIcon,
+  Cog6ToothIcon,
+  HomeIcon,
+  ShoppingCartIcon,
+  XMarkIcon,
+} from '@heroicons/react/24/outline';
+import { Fragment, useState } from 'react';
 
-const style = {
-  container: 'text-center h-screen bg-white pt-96',
-  text: 'inline-block align-middle text-5xl',
-};
+const navigation = [
+  { name: 'Dashboard', href: '#', icon: HomeIcon, current: true },
+  { name: 'Create', href: '/create', icon: CloudArrowUpIcon, current: false },
+  {
+    name: 'Buy/Sell',
+    href: '/collections',
+    icon: ShoppingCartIcon,
+    current: false,
+  },
+  { name: 'Performance', href: '#', icon: ChartBarIcon, current: false },
+  { name: 'Settings', href: '#', icon: Cog6ToothIcon, current: false },
+];
 
-const account = () => {
+function classNames(...classes: any) {
+  return classes.filter(Boolean).join(' ');
+}
+
+export default function Example() {
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+
   return (
-    <div className={style.container}>
-      <span className={style.text}>Dashboard</span>
-    </div>
-  );
-};
+    <>
+      {/*
+        This example requires updating your template:
 
-export default account;
+        ```
+        <html class="h-full bg-gray-100">
+        <body class="h-full">
+        ```
+      */}
+      <div>
+        <Transition.Root show={sidebarOpen} as={Fragment}>
+          <Dialog
+            as="div"
+            className="relative z-40 md:hidden"
+            onClose={setSidebarOpen}
+          >
+            <Transition.Child
+              as={Fragment}
+              enter="transition-opacity ease-linear duration-300"
+              enterFrom="opacity-0"
+              enterTo="opacity-100"
+              leave="transition-opacity ease-linear duration-300"
+              leaveFrom="opacity-100"
+              leaveTo="opacity-0"
+            >
+              <div className="fixed inset-0 bg-gray-600 bg-opacity-75" />
+            </Transition.Child>
+
+            <div className="fixed inset-0 z-40 flex">
+              <Transition.Child
+                as={Fragment}
+                enter="transition ease-in-out duration-300 transform"
+                enterFrom="-translate-x-full"
+                enterTo="translate-x-0"
+                leave="transition ease-in-out duration-300 transform"
+                leaveFrom="translate-x-0"
+                leaveTo="-translate-x-full"
+              >
+                <Dialog.Panel className="relative flex w-full max-w-xs flex-1 flex-col bg-indigo-700">
+                  <Transition.Child
+                    as={Fragment}
+                    enter="ease-in-out duration-300"
+                    enterFrom="opacity-0"
+                    enterTo="opacity-100"
+                    leave="ease-in-out duration-300"
+                    leaveFrom="opacity-100"
+                    leaveTo="opacity-0"
+                  >
+                    <div className="absolute top-0 right-0 -mr-12 pt-2">
+                      <button
+                        type="button"
+                        className="ml-1 flex h-10 w-10 items-center justify-center rounded-full focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
+                        onClick={() => setSidebarOpen(false)}
+                      >
+                        <span className="sr-only">Close sidebar</span>
+                        <XMarkIcon
+                          className="h-6 w-6 text-white"
+                          aria-hidden="true"
+                        />
+                      </button>
+                    </div>
+                  </Transition.Child>
+                  <div className="h-0 flex-1 overflow-y-auto pt-5 pb-4">
+                    <div className="flex flex-shrink-0 items-center px-4">
+                      <img
+                        className="h-8 w-auto"
+                        src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=300"
+                        alt="Your Company"
+                      />
+                    </div>
+                    <nav className="mt-5 space-y-1 px-2">
+                      {navigation.map((item) => (
+                        <a
+                          key={item.name}
+                          href={item.href}
+                          className={classNames(
+                            item.current
+                              ? 'bg-indigo-800 text-white'
+                              : 'text-white hover:bg-indigo-600 hover:bg-opacity-75',
+                            'group flex items-center px-2 py-2 text-base font-medium rounded-md'
+                          )}
+                        >
+                          <item.icon
+                            className="mr-4 h-6 w-6 flex-shrink-0 text-indigo-300"
+                            aria-hidden="true"
+                          />
+                          {item.name}
+                        </a>
+                      ))}
+                    </nav>
+                  </div>
+                  <div className="flex flex-shrink-0 border-t border-indigo-800 p-4">
+                    <a href="#" className="group block flex-shrink-0">
+                      <div className="flex items-center">
+                        <div>
+                          <img
+                            className="inline-block h-10 w-10 rounded-full"
+                            src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+                            alt=""
+                          />
+                        </div>
+                        <div className="ml-3">
+                          <p className="text-base font-medium text-white">
+                            Tom Cook
+                          </p>
+                          <p className="text-sm font-medium text-indigo-200 group-hover:text-white">
+                            View profile
+                          </p>
+                        </div>
+                      </div>
+                    </a>
+                  </div>
+                </Dialog.Panel>
+              </Transition.Child>
+              <div className="w-14 flex-shrink-0" aria-hidden="true">
+                {/* Force sidebar to shrink to fit close icon */}
+              </div>
+            </div>
+          </Dialog>
+        </Transition.Root>
+
+        {/* /********************************************************************************************************/
+        /*   STATIC SIDEBAR
+        /*******************************************************************************************************/}
+        <div className="hidden md:fixed md:inset-y-0 md:flex md:w-64 md:flex-col">
+          {/* Sidebar component, swap this element with another sidebar if you like */}
+          <div className="flex min-h-0 flex-1 flex-col bg-tblue">
+            <div className="flex flex-1 flex-col overflow-y-auto pt-5 pb-4">
+              <div className="flex flex-shrink-0 items-center px-4">
+                <img src="/assets/tokey_logo_white.svg" alt="Tokey" />
+              </div>
+              <nav className="mt-5 flex-1 space-y-1 px-2">
+                {navigation.map((item) => (
+                  <a
+                    key={item.name}
+                    href={item.href}
+                    className={classNames(
+                      item.current
+                        ? 'bg-tblue text-white'
+                        : 'text-white hover:bg-gray-300 hover:bg-opacity-75',
+                      'group flex items-center px-2 py-2 text-md font-medium rounded-md'
+                    )}
+                  >
+                    <item.icon
+                      className="mr-3 h-6 w-6 flex-shrink-0 text-white"
+                      aria-hidden="true"
+                    />
+                    {item.name}
+                  </a>
+                ))}
+              </nav>
+            </div>
+            {/* /********************************************************************************************************/
+            /*   PROFILE SECTION
+            /******************************************************************************************************/}
+            <div className="flex flex-shrink-0 border-t border-white p-4">
+              <a href="#" className="group block w-full flex-shrink-0">
+                <div className="flex items-center">
+                  <div>
+                    <img
+                      className="inline-block h-9 w-9 rounded-full"
+                      src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+                      alt=""
+                    />
+                  </div>
+                  <div className="ml-3">
+                    <p className="text-sm font-medium text-white">Tom Cook</p>
+                    <p className="text-xs font-medium text-indigo-200 group-hover:text-white">
+                      View profile
+                    </p>
+                  </div>
+                </div>
+              </a>
+            </div>
+          </div>
+        </div>
+        <div className="flex flex-1 flex-col md:pl-64">
+          <div className="sticky top-0 z-10 bg-gray-100 pl-1 pt-1 sm:pl-3 sm:pt-3 md:hidden">
+            <button
+              type="button"
+              className="-ml-0.5 -mt-0.5 inline-flex h-12 w-12 items-center justify-center rounded-md text-gray-500 hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500"
+              onClick={() => setSidebarOpen(true)}
+            >
+              <span className="sr-only">Open sidebar</span>
+              <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+            </button>
+          </div>
+          <main className="flex-1">
+            <div className="py-6">
+              <div className="mx-auto max-w-7xl px-4 sm:px-6 md:px-8">
+                <h1 className="text-2xl font-semibold text-gray-900">
+                  Dashboard
+                </h1>
+              </div>
+              <div className="mx-auto max-w-7xl px-4 sm:px-6 md:px-8">
+                {/* Replace with your content */}
+                <div className="py-4">
+                  <div className="h-96 rounded-lg border-4 border-dashed border-gray-200" />
+                </div>
+                {/* /End replace */}
+              </div>
+            </div>
+          </main>
+        </div>
+      </div>
+    </>
+  );
+}
